requirements.txt
appdirs==1.4.4
asgiref==3.5.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
cssselect==1.1.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-allauth==0.48.0
django-crispy-forms==1.13.0
django-filter==2.4.0
django-userforeignkey==0.4.0
djangorestframework==3.12.4
docker==5.0.3
fake-useragent==0.1.11
filelock==3.0.12
idna==2.10
lxml==4.6.3
mysqlclient==2.1.0
oauthlib==3.2.0
parse==1.19.0
Pillow==9.0.0
pycparser==2.21
pydub==0.25.1
pyee==8.1.0
PyJWT==1.7.1
pyppeteer==0.2.5
pyquery==1.4.3
python3-openid==3.2.0
pytz==2021.3
pywin32==227
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tqdm==4.61.1
urllib3==1.26.5
virtualenv==20.4.3
w3lib==1.22.0
webapi==0.1
websocket-client==1.3.1
websockets==8.1
whitenoise==5.0.1
